+++ CommonConcepts/Plugins/Rhetos.Dsl.DefaultConcepts/Computations/ComposableFilterByReferencedInfo.cs
@@ -50,12 +50,12 @@ namespace Rhetos.Dsl.DefaultConcepts
                     + "' must use a reference property that is a member of it's own data structure. Try using FilterByLinkedItems instead.");
 
             var availableFilters = concepts.FindByReference<ComposableFilterByInfo>(f => f.Source, ReferenceFromMe.Referenced)
-                .Select(f => f.Parameter).ToList();
+                .Select(f => f.Parameter).OrderBy(f => f).ToList();
 
             if (!availableFilters.Contains(Parameter))
                 throw new DslSyntaxException(this, string.Format(
                     "There is no {0} '{1}' on {2}. Available {0} filters are: {3}.",
-                    new ComposableFilterByInfo().GetKeywordOrTypeName(),
+                    ConceptInfoHelper.GetKeywordOrTypeName(typeof(ComposableFilterByInfo)),
                     Parameter,
                     ReferenceFromMe.Referenced.GetUserDescription(),
                     string.Join(", ", availableFilters.Select(parameter => "'" + parameter + "'"))));